1.breaking change(SDK directory structure):Change the directory structure to adapt to the multi-platform (826x/8258/8278/b91) SDK; 2.feat(b91):Support 9518 A0 chip; 3.feat(aps layer fragment):Aps layer supports zigbee fragment messages; 4.refactor(af layer):Strip the AF layer from the ZDO folder; 5.refactor(/proj):Clean up the abstraction layer driver.

+++ tl_zigbee_sdk/zigbee/zbhci/zbhci.h
@@ -311,7 +311,7 @@ typedef struct{
 	u8	srcEp;
 	u8	dstEp;
 	u16 clusterId;
-	u8  dataLen;
+	u16 dataLen;
 	u8  payload[];
 }zbhci_afDataSend_req_t;
 
@@ -320,7 +320,7 @@ typedef struct{
 	u8	srcEp;
 	u8	dstEp;
 	u16 clusterId;
-	u8  dataLen;
+	u16 dataLen;
 	u8  payload[];
 }zbhci_afDataSend_rsp_t;
